backend/requirements_prod.txt
# Core
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# API & Documentation
drf-spectacular==0.26.3
django-filter==23.3

# Authentication
djangorestframework-simplejwt==5.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
django-cors-headers==4.2.0

# Development (optional)
django-extensions==3.2.3